pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "mcp-auto-pr"
version = "0.2.0"
description = "MCP Auto PR - Orchestration and deployment for MCP PR recommendation system"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    {include = "mcp_local_repo_analyzer", from = "src"},
    {include = "mcp_pr_recommender", from = "src"},
    {include = "shared", from = "src"},
]

[tool.poetry.scripts]
mcp-local-analyzer = "mcp_local_repo_analyzer.main:main"
mcp-pr-recommender = "mcp_pr_recommender.main:main"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastmcp = "^2.10.6"
gitpython = "^3.1.40"
pydantic = {extras = ["mypy"], version = "^2.11.7"}
typing-extensions = "^4.8.0"
pathlib = "^1.0.1"
click = "^8.1.7"
fastapi = "^0.116.1"
sse-starlette = "^2.4.1"
uvicorn = "^0.35.0"
pydantic-settings = "^2.10.1"
openai = "^1.54.3"
httpx = "^0.28.1"
pyyaml = "^6.0.2"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-xdist = "^3.5.0"
pytest-benchmark = "^4.0.0"
pytest-timeout = "^2.2.0"
pytest-randomly = "^3.15.0"
pytest-clarity = "^1.0.1"
factory-boy = "^3.3.0"
freezegun = "^1.2.2"
responses = "^0.24.1"
hypothesis = "^6.92.1"
aioresponses = "^0.7.4"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
ruff = "^0.1.11"
mypy = "^1.8.0"
bandit = "^1.7.5"
safety = "^2.3.5"
commitizen = "^3.13.0"
interrogate = "^1.5.0"
types-pyyaml = "^6.0.12.20250516"
types-requests = "^2.31.0"
pydocstyle = "^6.3.0"
python-semantic-release = "^10.3.1"

# Black configuration removed - now using Ruff for formatting

[tool.ruff]
target-version = "py310"
line-length = 120
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort (import sorting)
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by ruff-format
]

[tool.ruff.isort]
known-first-party = ["shared", "mcp_local_repo_analyzer", "mcp_pr_recommender"]

# isort configuration removed - now using Ruff for import sorting

# MyPy configuration is in mypy.ini for better exclusion control

[tool.pytest.ini_options]
markers = [
    "unit: Fast unit tests",
    "integration: Integration tests requiring services",
    "slow: Tests that take >5 seconds",
    "external: Tests requiring external services (network, etc.)",
    "git: Tests requiring git repository setup",
    "benchmark: Performance benchmark tests",
]
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml"
]
